[SIX-21] feat : 별점 도메인을 삭제하고 요구사항에 맞게 리뷰 도메인 수정
