+++ onedayhero-domain/src/main/java/com/sixheroes/onedayherodomain/missionreview/ReviewerType.java
@@ -1,0 +1,4 @@
+package com.sixheroes.onedayherodomain.missionreview;
+
+public enum ReviewerType {
+}
