--- onedayhero-domain/src/main/java/com/sixheroes/onedayherodomain/star/ReviewerType.java
@@ -1,4 +1,0 @@
-package com.sixheroes.onedayherodomain.star;
-
-public enum ReviewerType {
-}
