--- onedayhero-domain/src/main/java/com/sixheroes/onedayherodomain/star/Star.java
@@ -1,33 +1,0 @@
-package com.sixheroes.onedayherodomain.star;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "stars")
-@Entity
-public class Star {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(name = "mission_id", nullable = false)
-    private Long missionId;
-
-    @Column(name = "sender_id", nullable = false)
-    private Long senderId;
-
-    @Column(name = "receiver_id", nullable = false)
-    private Long receiverId;
-
-    @Column(name = "score", nullable = false)
-    private Double score;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "reviewer_type", nullable = false)
-    private ReviewerType reviewerType;
-}
