--- onedayhero-domain/src/main/java/com/sixheroes/onedayherodomain/star/repository/StarRepository.java
@@ -1,7 +1,0 @@
-package com.sixheroes.onedayherodomain.star.repository;
-
-import com.sixheroes.onedayherodomain.star.Star;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StarRepository extends JpaRepository<Star, Long> {
-}
